// jobboard/frontend/lib/api/generated/types.gen.ts
export type CompanyPublic = {
    /**
     * Id
     */
    id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Slug
     */
    slug: string;
    /**
     * Legal Name
     */
    legal_name?: string | null;
    /**
     * Logo Url
     */
    logo_url?: string | null;
    /**
     * Website Url
     */
    website_url?: string | null;
    /**
     * Careers Page Url
     */
    careers_page_url?: string | null;
    /**
     * Short Description
     */
    short_description?: string | null;
    /**
     * Full Description
     */
    full_description?: string | null;
    /**
     * Industry
     */
    industry?: string | null;
    company_size_band?: CompanySizeBand | null;
    /**
     * Is Cdtm Startup
     */
    is_cdtm_startup?: boolean;
    /**
     * Hq City
     */
    hq_city?: string | null;
    /**
     * Hq Region
     */
    hq_region?: string | null;
    /**
     * Hq Country
     */
    hq_country?: string | null;
    /**
     * Linkedin Url
     */
    linkedin_url?: string | null;
    /**
     * Twitter Url
     */
    twitter_url?: string | null;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};

/**
 * CompanySizeBand
 */
export type CompanySizeBand = 'startup' | 'smb' | 'mid' | 'enterprise';